refactor: extract pipeline/output.rs
Move per-client output path computation (output_path,
skill/rule/agent_output_path) and the corresponding file I/O helpers
(write_output, remove_item_outputs) out of pipeline/mod.rs into a
focused submodule. Inline tests move with the functions.

output_path is no longer reachable by callers outside the pipeline
crate, so its visibility tightens from pub(crate) to pub(super) —
the only callers were inside the pipeline itself.

Refs #180
